Added Utility Code

+++ src/CoalFactory.cpp
@@ -1,5 +1,10 @@
 #include "CoalFactory.h"
 
-EnergySource* CoalFactory::createEnergySource(int output) {
+/**
+ * @brief Create an EnergySource object of type CoalSource
+ * 
+ * @param output The output of the CoalSource
+ */
+EnergySource* CoalFactory::createEnergySource(float output) {
 	return new CoalSource(output);
 }
