+++ src/EnergyFactory.h
@@ -2,9 +2,14 @@
 #define ENERGYFACTORY_H
 
 #include "EnergySource.h"
+/**
+ * @file EnergyFactory.h
+ * 
+ * @brief The header file for the EnergyFactory class
+ */
 class EnergyFactory {
 public:
-	virtual EnergySource* createEnergySource(int output) = 0;
+	virtual EnergySource* createEnergySource(float output) = 0;
 };
 
 #endif
