+++ src/NuclearFactory.h
@@ -3,9 +3,14 @@
 
 #include "EnergyFactory.h"
 #include "NuclearSource.h"
+/**
+ * @file NuclearFactory.h
+ * 
+ * @brief The header file for the NuclearFactory class
+ */
 class NuclearFactory : public EnergyFactory {
 public:
-	EnergySource* createEnergySource(int output);
+	EnergySource* createEnergySource(float output);
 };
 
 #endif
