+++ src/WindFactory.cpp
@@ -1,5 +1,10 @@
 #include "WindFactory.h"
 
-EnergySource* WindFactory::createEnergySource(int output) {
+/**
+ * @brief Create an EnergySource object of type WindSource
+ * 
+ * @param output The output of the WindSource
+ */
+EnergySource* WindFactory::createEnergySource(float output) {
 	return new WindSource(output);
 }
