+++ src/WindSource.h
@@ -2,9 +2,15 @@
 #define WINDSOURCE_H
 
 #include "EnergySource.h"
+/**
+ * @file WindSource.h
+ * 
+ * 
+ * @brief The header file for the WindSource class
+ */
 class WindSource : public EnergySource {
 public:
-	WindSource(int output);
+	WindSource(float output);
 
 	float getEnergyOutput();
